Fixes where time delta is calculated
